refactored `qmsmt` to adapt the new changes in acadia
This is currently a preliminary version which requires minimal changes to the old runtimes. I'm considering a larger refactoring. Currently the only changes are
1. in yaml file, `fixed_length` should be replaced by `stretch_length` to represent the number of 4 middle sample points that will be repeated.
2. prepare_cmacc should be removed now, and call resonator.measure directly with the window name
3. `InputOutput` now has a `pulse` function for playing pulses with extra stretch length, as an interface for writing pulse length sweeps with a register.
NOTE: Old codes that call `schedule_waveform` with a WaveformMemory as input might need to be refactored, as the `fixed_length` is no longer an attribute of WaveformMemory, and stretch_length must be supplied manually.
4. John added `load_multichromatic_waveform`, which can be used for playing multiplexed  pulses within the sampling bandwidth of the DMA (~+- 300MHz) from the same DAC channel.

Co-Authored-By: John Garmon <john.garmon@yale.edu>